testdata: rename testdata to testdatasource to circumvent convention in go
Before this change the tests was not run.
Seems like angular directive inputDatetime is no longer there so
had to make a fix to be able to get the manual input scenario to
work.
