+++ pkg/tsdb/testdatasource/utils.go
@@ -1,4 +1,4 @@
-package testdata
+package testdatasource
 
 import (
 	"math/rand"
