+++ public/app/plugins/datasource/testdata/partials/query.editor.html
@@ -35,7 +35,7 @@
 			<label class="gf-form-label query-keyword width-7">New value</label>
       <input type="number" class="gf-form-input width-15" placeholder="value" ng-model="ctrl.newPointValue">
 			<label class="gf-form-label query-keyword">Time</label>
-      <input type="string" class="gf-form-input width-12" placeholder="time" ng-model="ctrl.newPointTime" input-datetime>
+      <input type="string" class="gf-form-input width-12" placeholder="time" ng-model="ctrl.newPointTime">
 			<button class="btn btn-secondary gf-form-btn" ng-click="ctrl.addPoint()">Add</button>
 			<label class="gf-form-label query-keyword">All values</label>
 			<gf-form-dropdown css-class="width-12" model="ctrl.selectedPoint" get-options="ctrl.getPoints()" on-change="ctrl.pointSelected($option)">
